requirements.txt
# ─── Web Framework ────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.29.0

# ─── Scheduler ────────────────────────────────────────────
apscheduler==3.10.4

# ─── Telegram ─────────────────────────────────────────────
python-telegram-bot==21.0.1

# ─── Email ────────────────────────────────────────────────
aiosmtplib==3.0.1

# ─── Database ─────────────────────────────────────────────
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.29

# ─── HTTP Client ──────────────────────────────────────────
httpx==0.27.0

# ─── Data Validation ──────────────────────────────────────
pydantic==2.7.0
pydantic-settings==2.2.1

# ─── Auth ─────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ─── Dev Tools ────────────────────────────────────────────
black==24.4.0
isort==5.13.2
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
